move zero padding back
