+++ src/layer/conv.rs
@@ -4,6 +4,7 @@ use crate::layer::Layer;
 use crate::onnx;
 use crate::util;
 use ark_bn254::Fr;
+use copy_constraint::zero_padding_partition;
 use ndarray::indices;
 use ndarray::Dim;
 use ndarray::Dimension;
@@ -125,7 +126,7 @@ impl Layer for ConvLayer {
     let permutation = splat_input(&input_shapes[0], &strides, &pads, weight_shape[1], &ch_dims);
     let permutation_padded = splat_pad(&permutation);
     let input_shape_padded: Vec<_> = input_shapes[0].iter().map(|i| i.next_power_of_two()).collect();
-    let padding_partitions = util::zero_padding_partition(&permutation_padded);
+    let padding_partitions = zero_padding_partition(&permutation_padded);
     let cc = graph.addBB(Box::new(CopyConstraintBasicBlock {
       permutation: permutation_padded,
       input_dim: IxDyn(&input_shape_padded),
@@ -136,7 +137,7 @@ impl Layer for ConvLayer {
     let weights_splatted = splat_weights(&weight_shape);
     let weights_padded = splat_pad(&weights_splatted);
     let weight_shape_padded: Vec<_> = weight_shape.iter().map(|i| i.next_power_of_two()).collect();
-    let padding_partitions = util::zero_padding_partition(&weights_padded);
+    let padding_partitions = zero_padding_partition(&weights_padded);
     let cc1 = graph.addBB(Box::new(CopyConstraintBasicBlock {
       permutation: weights_padded,
       input_dim: IxDyn(&weight_shape_padded),
@@ -170,7 +171,7 @@ impl Layer for ConvLayer {
     let mut output_shape = vec![1, weight_shape[0]];
     output_shape.append(&mut out_dims);
     let reshape_permutation = reshape_permutation(&vec![permutation.len(), weights_splatted.len()], &output_shape);
-    let padding_partitions = util::zero_padding_partition(&reshape_permutation);
+    let padding_partitions = zero_padding_partition(&reshape_permutation);
     let cc2 = graph.addBB(Box::new(CopyConstraintBasicBlock {
       permutation: reshape_permutation,
       input_dim: IxDyn(&[permutation.len().next_power_of_two(), weights_splatted.len().next_power_of_two()]),
