+++ src/layer/flatten.rs
@@ -3,6 +3,7 @@ use crate::graph::*;
 use crate::layer::Layer;
 use crate::util;
 use ark_bn254::Fr;
+use copy_constraint::zero_padding_partition;
 use ndarray::{ArrayD, IxDyn};
 use tract_onnx::pb::AttributeProto;
 
@@ -36,7 +37,7 @@ impl Layer for FlattenLayer {
 
     let (permutation, output_shape) = get_permutation(&input_shapes[0], axis);
 
-    let padding_partitions = util::zero_padding_partition(&permutation);
+    let padding_partitions = zero_padding_partition(&permutation);
     let cc = graph.addBB(Box::new(CopyConstraintBasicBlock {
       permutation: permutation,
       input_dim: IxDyn(&padded_input_shape),
