+++ src/util/copy_constraint.rs
@@ -45,21 +45,6 @@ pub fn max_padding_partitions(permutation: &ArrayD<Option<IxDyn>>, val: Fr) -> H
   partitions
 }
 
-// Returns the padding_partitions field for CopyConstraintBasicBlock when the given permutation padding elements are 0
-pub fn zero_padding_partition(permutation: &ArrayD<Option<IxDyn>>) -> HashMap<Fr, Vec<IxDyn>> {
-  let mut partition = vec![];
-  for (i, _) in permutation.indexed_iter() {
-    if permutation[&i] == None {
-      partition.push(i);
-    }
-  }
-  let mut padding_partition = HashMap::new();
-  if partition.len() > 0 {
-    padding_partition.insert(Fr::zero(), partition);
-  }
-  padding_partition
-}
-
 // Helper function to get the indices of the reshaped tensor
 // Note that the input_shape and output_shape are non-padded
 pub fn get_reshape_indices(input_shape: Vec<usize>, output_shape: Vec<usize>) -> ArrayD<Option<IxDyn>> {
